Default vertex color 1.0 1.0 1.0

+++ nvl/utilities/color.h
@@ -71,10 +71,10 @@ public:
 
 private:
 
-    int vRed;
-    int vGreen;
-    int vBlue;
-    int vAlpha;
+    float vRed;
+    float vGreen;
+    float vBlue;
+    float vAlpha;
 
 };
